+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
+
+## [1.2.0](https://github.com/arangodb/go-driver/tree/1.2.0) (2021-08-04)
 - Add support for AQL, Pipeline, Stopwords, GeoJSON and GeoPoint Arango Search analyzers.
 - Add `estimates` field to indexes properties.
 - Add tests for 3.8 ArangoDB and remove tests for 3.5.